Issue #10: rake task to ingest clip Print proper usage; Function to create tar removed; Command line parameters added for storage_machine_id and capture_machine_id; Commented code to send the tar files are also removed; When run in production mode, file in directory to store clips and thumbnails based on ids is deleted

+++ lib/tasks/ingest_clip.rake
@@ -5,8 +5,6 @@ require 'open3'
 
 namespace :ingest_clip do
 
-# rake ingest_clip:ingest stream_id=<ID> clip_file=<clip_file_location>
-
   def fail_on_error(wait_thr, cmd)
     exit_status = wait_thr.value
     unless exit_status.success?
@@ -35,68 +33,59 @@ namespace :ingest_clip do
   end
 
   def clip_info( clip_file)
-    op_str = execute_and_read( "ffprobe -v quiet -print_format json  -show_streams #{clip_file}")
-    jsObj  = JSON.parse( op_str)
+    op_str  = execute_and_read( "ffprobe -v quiet -print_format json  -show_streams #{clip_file}")
+    jsObj   = JSON.parse( op_str)
     streams = jsObj['streams']
 
-    info = Hash.new
+    info    = Hash.new
     streams.each { |strm_obj|
       strm_obj.each { |k,v| info[k] = v }
     }
     info
   end
 
-  def buildTar(data_folder, tarname)
-    puts("Data: #{data_folder} Tar: #{tarname}")
-    Dir.chdir( data_folder) do
-      tar_cmd = "tar -zcvf #{tarname} *"
-      execute_cmd tar_cmd
-    end
-    tarname
-  end
-
 
   desc "Setup the stream id and video folder file"
   task setup: :environment do 
-    @is_production =  "production" == ENV["RENV"]
-    @stream_id = ENV['stream_id'] 
-    @clip_file = ENV['clip_file']
+    @is_production      =  "production" == ENV["RENV"]
+    @stream_id          = ENV['stream_id'] 
+    @clip_file          = ENV['clip_file']
+    @storage_machine_id = ENV['storage_machine_id'] 
+    @capture_machine_id = ENV['capture_machine_id'] 
 
-    @storage_machine_id = 1
-    @capture_machine_id = 2
-    @gpu_machind_id =3
 
     puts "Stream #{@stream_id} Clip: #{@clip_file}"
-    if @stream_id.nil? || @clip_file.nil?
+    if @stream_id.nil? || @clip_file.nil? || @storage_machine_id.nil? || @capture_machine_id.nil?
+      puts "Usage: rake ingest_clip:ingest stream_id=<stream-id> clilp_file=<clip-file> storage_machine_id=<storage-machine-id> capture_machine_id=<caputre-machine-id>"
       abort "Both stream_id and clip_file must be defined environment variables"
     end
 
     if ! File.exists? @clip_file
       abort "Can not find or access the file #(@clip_file)"
     end
-    @clip_dir = File.dirname( @clip_file)
-    @file_name = File.basename( @clip_file, ".*")
+    @clip_dir        = File.dirname( @clip_file)
+    @file_name       = File.basename( @clip_file, ".*")
     @tmp_folder_root = "/tmp/task_process" #{Process.pid}
 
   end
 
   def buildCaptureModel(streamModel, clip_dim)
-    captureModel = Video::Capture.create()
-    captureModel.stream = streamModel
+    captureModel                    = Video::Capture.create()
+    captureModel.stream             = streamModel
     captureModel.storage_machine_id = @storage_machine_id
     captureModel.capture_machine_id = @capture_machine_id
-    captureModel.width = clip_dim['width']
-    captureModel.height = clip_dim['height']
+    captureModel.width              = clip_dim['width']
+    captureModel.height             = clip_dim['height']
     captureModel.save
     captureModel
   end
 
   def buildClip( captureModel)
-    clip = Video::Clip.create()
-    clip.capture = captureModel
-    clip.length = 60000 # 60 * 1000
+    clip                    = Video::Clip.create()
+    clip.capture            = captureModel
+    clip.length             = 60000 # 60 * 1000
     clip.frame_number_start = 0
-    clip.frame_number_end = 1500 # 25 * 60
+    clip.frame_number_end   = 1500 # 25 * 60
     clip.save
     clip
   end
@@ -137,9 +126,12 @@ namespace :ingest_clip do
       puts("Status #{status} Trace #{trace} serverFilePath #{serverFilePath}")
 
       data_folder ="#{@tmp_folder_root}/data" 
-      rel_CaptureFolder = "streams/#{@stream_id}/#{captureModel.id}"
+      rel_StreamFolder = "streams/#{@stream_id}"
+      @localDataFolder = "#{serverFilePath}/#{rel_StreamFolder}"
+      rel_CaptureFolder = "#{rel_StreamFolder}/#{captureModel.id}"
+
+
       tmp_folder = "#{data_folder}/#{rel_CaptureFolder}"
-      FileUtils.mkdir_p tmp_folder
 
 
       if File.exists? tmp_folder
@@ -160,32 +152,27 @@ namespace :ingest_clip do
 
         puts "CLIP: \n#{clip.inspect} #{clip.id}"
         
-        clip_folder = @is_production ? "#{tmp_folder}/#{clip.id.to_s}" :  serverClipFolder
-        new_clip = "#{clip_folder}/#{clip.id.to_s}"
+        clip_folder     = @is_production ? "#{tmp_folder}/#{clip.id.to_s}" :  serverClipFolder
+        new_clip        = "#{clip_folder}/#{clip.id.to_s}"
 
 
         FileUtils.mkdir_p clip_folder
-        new_clip_name = "#{new_clip}#{File.extname(f)}"
+        new_clip_name   = "#{new_clip}#{File.extname(f)}"
         thumb_nail_file = "#{new_clip}.jpg"
         puts("new clip name #{new_clip_name} thumb namil #{thumb_nail_file}")
         FileUtils.mv( f, new_clip_name)
 
         
-        status, trace = storageClient.saveFile( new_clip_name, serverClipPath )
+        status, trace   = storageClient.saveFile( new_clip_name, serverClipPath )
         puts("Status send #{status} #{trace} for #{new_clip_name} TO: #{serverClipPath}")
 
-        thumbnail_cmd = "ffmpeg -y -i #{new_clip_name} -f mjpeg -vframes 1 #{thumb_nail_file}"
+        thumbnail_cmd   = "ffmpeg -y -i #{new_clip_name} -f mjpeg -vframes 1 #{thumb_nail_file}"
         execute_cmd thumbnail_cmd
 
-        status, trace = storageClient.saveFile( thumb_nail_file, serverThumbnailPath )
+        status, trace   = storageClient.saveFile( thumb_nail_file, serverThumbnailPath )
         puts("Status send #{status} #{trace} for #{thumb_nail_file} TO: #{serverThumbnailPath}")
       end
 
-      # tar_file = buildTar(data_folder, "#{tar_folder}/build_inputs.tar.gz")
-      # status, trace = storageClient.saveFile( tar_file, "#{serverFilePath}/build_inputs.tar.gz")
-      # puts("send #{data_folder} to #{serverFilePath}")
-      # status, trace = storageClient.saveFile( data_folder, serverFilePath)
-      # puts "File send status #{status} #{trace}"
       storageClient.closeConnection
     ensure
       Rake::Task["ingest_clip:cleanup"].execute
@@ -193,13 +180,16 @@ namespace :ingest_clip do
   end
 
   task :cleanup do
-    FileUtils.rm_r @tmp_folder_root
-    if @is_production
-      
+    if File.exists? @tmp_folder_root
+      FileUtils.rm_r @tmp_folder_root
+    end
+    puts("#{@localDataFolder} #{ File.exists? @localDataFolder}")
+    if @is_production && (File.exists? @localDataFolder)
+      puts "remove local data foler"
+      FileUtils.rm_r @localDataFolder
     end
   end
 
 end
-
